This is a larger commit that I ususally like to do
It changes the config system, such that the
config file that is loaded depends on the switch state.
For example if the switch state is 11, the config file
config11.ini is used.  I believe that this is the most
general way to have different running states, that
are also completly flexale.